Dish Validator

+++ borsch/src/main/java/com/nayax/borsch/validation/config/PageIdValidationConfig.java
@@ -1,11 +1,13 @@
 package com.nayax.borsch.validation.config;
 
 import com.nayax.borsch.model.dto.assortment.request.ReqSimplePriceItemUpDto;
+import com.nayax.borsch.repository.impl.AdditionsRepository;
 import com.nayax.borsch.validation.Validator;
 import com.nayax.borsch.validation.ValidatorComponent;
 import com.nayax.borsch.validation.componentimpl.SimpleValidatorComponent;
 import com.nayax.borsch.validation.enums.ValidationAction;
 import com.nayax.borsch.validation.impl.ValidatorImpl;
+import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
 import java.util.Objects;
@@ -21,7 +23,7 @@ public class PageIdValidationConfig {
                 "Parameters of page are null"));
         validatorPageId.add(SimpleValidatorComponent.getComponents(List.of(ValidationAction.ADDITIONS_GETALL,
                 ValidationAction.DISH_GETALL,ValidationAction.DRINK_GETALL),obj-> obj != null
-                        && (Long)obj > 0 ,
+                        && (Integer)obj > 0 ,
                 "Parameters of page are invalid"));
         validatorPageId.add(SimpleValidatorComponent.getComponents(List.of(ValidationAction.ADDITIONS_GETALL,
                 ValidationAction.DISH_GETALL,ValidationAction.DRINK_GETALL), Objects::nonNull,
